chore(monomodule): refactors utils package.
